refactor - changed project structure
