+++ src/App.jsx
@@ -1,11 +1,11 @@
 import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Navbar from "./components/Navbar";
-import Dashboard from "./pages/Dashboard";
-import Ledger from "./pages/Ledger";
-import Reports from "./pages/Reports";
+import Navbar from "./components/shared/Navbar";
+import Dashboard from "./pages/dashboard/container/Dashboard";
+import Ledger from "./pages/ledger/container/Ledger";
+import Reports from "./pages/reports/container/Reports";
 import RouteConstants from "./common/RouteConstants";
-import Topbar from "./components/Topbar";
+import Topbar from "./components/shared/Topbar";
 
 export default function App() {
   return (
